Connect toolbar once instead of each function
