+++ ui/src/flux/components/flux_functions_toolbar/FluxFunctionsToolbar.tsx
@@ -1,5 +1,6 @@
 // Libraries
 import React, {PureComponent} from 'react'
+import {Subscribe} from 'unstated'
 
 // Components
 import TransformToolbarFunctions from 'src/flux/components/flux_functions_toolbar/TransformToolbarFunctions'
@@ -10,14 +11,22 @@ import FancyScrollbar from 'src/shared/components/FancyScrollbar'
 // Constants
 import {functions as FUNCTIONS} from 'src/flux/constants'
 
+// Utils
+import {TimeMachineContainer} from 'src/shared/utils/TimeMachineContainer'
+
 // Decorators
 import {ErrorHandling} from 'src/shared/decorators/errors'
+
+interface ConnectedProps {
+  script: string
+  onUpdateScript: (script: string) => void
+}
 
 interface State {
   searchTerm: string
 }
 @ErrorHandling
-class FluxFunctionsToolbar extends PureComponent<{}, State> {
+class FluxFunctionsToolbar extends PureComponent<ConnectedProps, State> {
   public constructor(props) {
     super(props)
     this.state = {searchTerm: ''}
@@ -41,6 +50,7 @@ class FluxFunctionsToolbar extends PureComponent<{}, State> {
                         key={category}
                         category={category}
                         funcs={funcs}
+                        onClickFunction={this.handleUpdateScript}
                       />
                     )
                   }
@@ -56,6 +66,24 @@ class FluxFunctionsToolbar extends PureComponent<{}, State> {
   private handleSearch = (searchTerm: string): void => {
     this.setState({searchTerm})
   }
+
+  private handleUpdateScript = (funcExample: string) => {
+    const {script, onUpdateScript} = this.props
+
+    const updatedScript = `${script}\n  |> ${funcExample}`
+    onUpdateScript(updatedScript)
+  }
 }
 
-export default FluxFunctionsToolbar
+const ConnectedFluxFunctionsToolbar = () => (
+  <Subscribe to={[TimeMachineContainer]}>
+    {(container: TimeMachineContainer) => (
+      <FluxFunctionsToolbar
+        script={container.state.draftScript}
+        onUpdateScript={container.handleUpdateDraftScript}
+      />
+    )}
+  </Subscribe>
+)
+
+export default ConnectedFluxFunctionsToolbar
